cut down old code and add measurements
